LayoutTests/http/tests/inspector/network/network-disable-cache-xhrs-expected.txt

CONSOLE MESSAGE: line 23: First XHR loaded.
CONSOLE MESSAGE: line 33: Done.
Tests disabling cache from inspector.

http://127.0.0.1:8000/inspector/network/resources/resource.php?random=1&cached=1
http://127.0.0.1:8000/inspector/network/resources/resource.php?random=1&cached=1
http://127.0.0.1:8000/inspector/network/resources/resource.php?random=1&cached=1
